@@ -3,9 +3,9 @@ import { useSelector, useDispatch } from 'react-redux'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 
-import Button from '@material-ui/core/Button'
-import IconButton from '@material-ui/core/IconButton'
-import Modal from '@material-ui/core/Modal'
+import Button from '@mui/material/Button'
+import IconButton from '@mui/material/IconButton'
+import Modal from '@mui/material/Modal'
 
 import { IconFont } from '../icons'
 import Selector from '../selector'
@@ -98,6 +98,8 @@ const canvasCtx = {
   pattern: null,
   preview: null,
 }
+
+const TextEditorRef = React.forwardRef((props, ref) => <TextEditor {...props} forwardedRef={ref} />)
 
 const Painter = ({screen}) => {
   const dispatch = useDispatch()
@@ -247,8 +249,9 @@ const Painter = ({screen}) => {
   }
 
   const handleMouseTouchDown = (clientX, clientY) => {
+    const { x, y } = stretcherRef.current.getBoundingClientRect()
     setIsMouseDown(true)
-    handleStretcherPosiion(clientX, clientY)
+    handleStretcherPosiion(clientX - x, clientY - y)
   }
 
   const handleMouseTouchUp = () => {
@@ -515,15 +518,12 @@ const Painter = ({screen}) => {
                 key={d.val}
                 onClick={ () => setTool(d.val) }
                 color={ tool === d.val ? 'primary' : 'default' }
-              >
+                size="large">
                 <IconFont style={d.icon} />
               </IconButton>
             ))
           }
-          <IconButton
-            disabled={ step < 1 }
-            onClick={backToPreviousStorage}
-          >
+          <IconButton disabled={ step < 1 } onClick={backToPreviousStorage} size="large">
             <IconFont style="step-back" />
           </IconButton>
         </div>
@@ -581,9 +581,7 @@ const Painter = ({screen}) => {
             {
               tool === 'text' && (
                 <div className="painter__text-tool">
-                  <IconButton
-                    onClick={() => setTextEditorOpen(true)}
-                  >
+                  <IconButton onClick={() => setTextEditorOpen(true)} size="large">
                     <IconFont style="text-edit" />
                   </IconButton>
                   <div className="painter__text-patterns">
@@ -592,7 +590,7 @@ const Painter = ({screen}) => {
                         <IconButton
                           onClick={ () => setTextPatternIndex(i) }
                           color={ textPatternIndex === i ? 'primary' : 'default' }
-                        >
+                          size="large">
                           <TextPattern pattern={d} size={28} />
                         </IconButton>
                       ))                      
@@ -610,12 +608,12 @@ const Painter = ({screen}) => {
         className="editor__modal"
         keepMounted={true}
       >
-        <TextEditor
+        <TextEditorRef
           onClose={handleTextEditorClose}
         />
       </Modal>
     </div>
-  )
+  );
 }
 
 Painter.propTypes = {}
