@@ -2,32 +2,11 @@ const webpack = require('webpack')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 
-const htmlTemplate = new HtmlWebpackPlugin({
-  template: `index.ejs`,
-})
-
-const devServer = {
-  inline: true,
-  https: false,
-  open: true,
-  host: '0.0.0.0',
-  port: 3030,
-  historyApiFallback: true,
-}
-
-const defaultSetting = {
-  page: 'index',
-  noMockServer: false,
-}
-
 module.exports = (env, argv) => {
-  const { page } = Object.assign(defaultSetting, argv)
-
-  const webpackSetting = {
-    mode: env === 'dev' ? 'development' : 'production',
+  return ({
     context: `${__dirname}/`,
     entry: {
-      web: './index.js',
+      web: './src/index.jsx',
     },
     output: {
       path: `${__dirname}/dist/`,
@@ -50,13 +29,11 @@ module.exports = (env, argv) => {
         },
         {
           test: /\.(gif|png|jpe?g)$/i,
+          type: 'asset/resource',
+          generator: {
+            filename: '[path][name].[ext]'
+          },
           use: [
-            {
-              loader: 'file-loader',
-              options: {
-                name: '[path][name].[ext]',
-              },
-            },
             {
               loader: 'image-webpack-loader',
               options: {
@@ -79,24 +56,22 @@ module.exports = (env, argv) => {
         },
         {
           test: /\.(eot|svg|ttf|woff|woff2)$/,
-          use: 'file-loader?name=public/fonts/[name].[ext]',
+          type: 'asset/resource',
+          generator: {
+            filename: 'public/fonts/[name].[ext]'
+          }
         },
         { test: /\.html$/, use: 'html-loader' },
       ],
     },
     plugins: [
-      htmlTemplate,
+      new HtmlWebpackPlugin({
+        template: `src/index.ejs`,
+      }),
       new MiniCssExtractPlugin(),
     ],
     resolve: {
       extensions: ['.js', '.jsx'],
     },
-  }
-
-  if (env === 'dev') {
-    webpackSetting['devtool'] = 'source-map'
-    webpackSetting['devServer'] = devServer
-  }
-
-  return webpackSetting
+  })
 }
